+++ app/src/main/java/com/example/myapplication/parking/VehicleParkedActivity.java
@@ -84,4 +84,5 @@ public class VehicleParkedActivity extends AppCompatActivity {
         startActivity(myIntent);
         return true;
     }
+
 }
