+++ app/src/main/res/values/colors.xml
@@ -9,4 +9,5 @@
     <color name="white">#FFFFFFFF</color>
     <color name="red_500">#B50D0D</color>
     <color name="red_700">#D32F2F</color>
+    <color name ="grey">#5A5A5A</color>
 </resources>
